cnp_mcmc

cnp_mcmc

Usage

cnp_mcmc(TL, param, iter, params_st, cor, ...)

Arguments

TL

Total length(s) in cm

param

List of all parameter means (add "_m") and standard deviations (add "_sd") Default parameters are set with very low sd's. parameters:

Qc_m, Qc_sd: percentage C of dry mass fish

Qn_m, Qn_sd: percentage N of dry mass fish

Qp_m, Qp_sd: percentage P of dry mass fish

Dc_m, Dc_sd: percentage C of dry mass food

Dn_m, Dn_sd: percentage N of dry mass food

Dp_m, Dp_sd: percentage P of dry mass food

ac_m, ac_sd: C-specific assimilation efficiency

an_m, an_sd: N-specific assimilation efficiency

ap_m, ap_sd: P-specific assimilation efficiency

linf_m, linf_sd: Von Bertalanffy Growth parameter, theoretical maximum size in TL (cm)

k_m, k_sd: Von Bertalanffy Growth parameter, growth rate (yr^-1)

t0_m, tO_sd: Von Bertalanffy Growth parameter (yr)

lwa_m, lwa_sd: Parameter length-weight relationship (g cm^-1)

lwb_m, lwb_sd: Parameter length-weight relationship

mdw_m, wprop_sd: Ratio between dry weight and wet weight of fish

F0nz_m, F0nz_sd: N-specific turnover rate

F0pz_m, F0pz_sd: P-specific turnover rate

f0_m, f0_sd: Metabolic normalisation constant independent of body mass (g C g^-alpha d^-1)

alpha_m, alpha_sd: Metabolic rate mass-scaling exponent

theta_m, theta_sd: Activity scope

r_m, r_sd: Aspect ratio of caudal fin

h_m, h_sd: Trophic level

v_m, v_sd: Environmental temperature (degrees celcius)

iter

A positive integer specifying the number of iterations. The default is 2000.

params_st

Standard parameters.

cor

A list of correlations between certain parameters: ro_Qc_Qn, ro_Qc_Qp, ro_Qn_Qp, ro_Dc_Dn, ro_Dc_Dp, ro_Dn_Dp, ro_lwa_lwb, ro_alpha_f0

...

Additional arguments rstan::sampling, see ?rstan:sampling