A function to extract specific model output parameters from result

A function to extract specific model output parameters from result

extract(mod, par)

Arguments

mod

Output from cnp_mod_mcmc()

par

Character vector specifying which output parameter that should be returned.

Value

Main model output parameters:

F0c: C-specific minimal inorganic flux (g/day)

F0n: N-specific minimal inorganic flux (g/day)

F0p: P-specific minimal inorganic flux (g/day)

Gc: Carbon-specific growth rate (g/day)

Gn: Nitrogen-specific growth rate (g/day)

Gp: Phosphorus-specific growth rate (g/day)

Sc: C-specific minimal supply rate (g/day)

Sn: N-specific minimal supply rate (g/day)

Sp: P-specific minimal supply rate (g/day)

Ic: Ingestion rate of C (g/day)

In: Ingestion rate of N (g/day)

Ip: Ingestion rate of P (g/day)

Wc: Egestion rate of C (g/day)

Wn: Egestion rate of N (g/day)

Wp: Egestion rate of P (g/day)

Fc: Total inorganic flux of C (respiration) (g/day)

Fn: Total inorganic flux of N (excretion) (g/day)

Fp: Total inorganic flux of P (excretion) (g/day)

Returns a data.frame with a summary of the selected output parameters

Examples

model <- cnp_model_mcmc(TL = 5:10, param = list(Qc_m = 40, Qn_m = 10, Qp_m = 4))
extract(model, c("Fn","Fp"))
#>   TL      Fn_mean    Fn_median        Fn_sd      Fn_2.5%     Fn_97.5%
#> 1  5 0.0006047886 0.0006047886 6.501450e-12 0.0006047886 0.0006047886
#> 2  6 0.0010610097 0.0010610097 1.081441e-11 0.0010610097 0.0010610097
#> 3  7 0.0017065393 0.0017065393 1.743413e-11 0.0017065393 0.0017065394
#> 4  8 0.0025757605 0.0025757605 2.482172e-11 0.0025757604 0.0025757605
#> 5  9 0.0037034689 0.0037034689 3.558761e-11 0.0037034688 0.0037034690
#> 6 10 0.0051248251 0.0051248251 5.172689e-11 0.0051248250 0.0051248252
#>         Fn_25%       Fn_75%      Fp_mean    Fp_median        Fp_sd      Fp_2.5%
#> 1 0.0006047886 0.0006047886 0.0001097965 0.0001097965 1.905610e-12 0.0001097965
#> 2 0.0010610097 0.0010610097 0.0002186548 0.0002186548 3.145278e-12 0.0002186548
#> 3 0.0017065393 0.0017065394 0.0003815567 0.0003815567 5.078797e-12 0.0003815567
#> 4 0.0025757605 0.0025757605 0.0006096826 0.0006096826 7.242953e-12 0.0006096826
#> 5 0.0037034689 0.0037034689 0.0009143614 0.0009143614 1.034990e-11 0.0009143614
#> 6 0.0051248250 0.0051248251 0.0013070515 0.0013070515 1.512917e-11 0.0013070515
#>       Fp_97.5%       Fp_25%       Fp_75%
#> 1 0.0001097965 0.0001097965 0.0001097965
#> 2 0.0002186548 0.0002186548 0.0002186548
#> 3 0.0003815567 0.0003815567 0.0003815567
#> 4 0.0006096827 0.0006096826 0.0006096826
#> 5 0.0009143614 0.0009143614 0.0009143614
#> 6 0.0013070515 0.0013070515 0.0013070515