A function to find length-weight relationship parameters a and b

A function to find estimates length-weight relationship parameters available on fishbase. It returns a list of means and standard deviations of a and b obtained from: Froese, R., J. Thorson and R.B. Reyes Jr., 2013. A Bayesian approach for estimating length-weight relationships in fishes. J. Appl. Ichthyol. (2013):1-7. Please cite Froese et al. (2013), when using these values. The default mirror for fishbase is set to "se", please change this if needed

Usage

find_lw(sp, mirror = "se")

Arguments

sp

A charachter value containing the species name

mirror

Mirror for fishbase (eg. "se", "org", "us", etc.) Default is

Value

A dataframe with means and standard deviations of length-weight parameters

Examples

# \donttest{
library(fishflux)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# find length-weight relationship parameters for one species
find_lw("Lutjanus griseus")
#>            species   lwa_m      lwa_sd lwb_m     lwb_sd
#> 1 Lutjanus griseus 0.01479 0.001030612  2.97 0.01020408

# find length-weight relationship parameters for multiple species and return
#   in data.frame
lapply(c("Chlorurus spilurus","Zebrasoma scopas"), find_lw) %>%
  dplyr::bind_rows()
#>              species   lwa_m      lwa_sd lwb_m     lwb_sd
#> 1 Chlorurus spilurus 0.01259 0.004918367  3.03 0.07653061
#> 2   Zebrasoma scopas 0.02512 0.002709184  2.98 0.02551020
# }