The 'fishflux' package.

The fishflux package provides a tool to model fluxes of C (carbon), N (nitrogen) and P (phosphorus) in fish. It combines basic principles from elemental stoichiometry and metabolic theory. The package offers a user-friendly interface to make nutrient dynamic modelling available for anyone. fishflux is mostly targeted towards fish ecologists, wishing to predict nutrient ingestion, egestion and excretion to study fluxes of nutrients and energy. Main assets:

Provides functions to model fluxes of Carbon, Nitrogen and Phosphorus for fish with or without the MCMC sampler provided by stan.

Provides some tools to find the right parameters as inputs into the model

Provides a plotting function to illustrate results

See also

Author

Maintainer: Nina M. D. Schiettekatte nina.schiettekatte@gmail.com

Authors:

Diego Barneche d.barneche@aims.gov.au