A function to find growth parameters on fishbase

A function to find growth parameters of a species using rfishbase. It returns a data frame containing K, t0 and Linf, the source. This function is useful to see what is available on fishbase. Nevertheless, we strongly recommend to check the source and only use otolith based studies.

growth_params(sp, otolith = TRUE)

Arguments

sp
A charachter value containing the species name

otolith
A logical value. If TRUE, only results from otolith analysis are returned. If false, all growth studies will be returned.

Examples

growth_params("Lutjanus griseus")

#> # A tibble: 6 x 7
#>   species        Locality             k  Linf     t0 method             comments
#>   <chr>          <chr>            <dbl> <dbl>  <dbl> <chr>              <chr>   
#> 1 Lutjanus gris… Southeast Cuba   0.24   51   -0.62  annuli on many ot… NA      
#> 2 Lutjanus gris… NA               0.228  55   -1.06  annuli on many ot… NA      
#> 3 Lutjanus gris… Southwest  Zone  0.15   50.3 -1.84  annuli on many ot… NA      
#> 4 Lutjanus gris… Florida (east c… 0.1    89   -0.32  annuli on many ot… NA      
#> 5 Lutjanus gris… South Florida    0.13   62.5 -1.33  annuli on many ot… NA      
#> 6 Lutjanus gris… North Florida    0.17   71.7 -0.025 annuli on many ot… NA