A function to find a set of parameters

A function to find a set of parameters

model_parameters(sp, family, otolith = TRUE, temp, ...)

Arguments

sp
Species name

family
family

otolith
TRUE or FALSE, if TRUE, function will only search fishbase for growth parameters that are based upon otolith analysis

temp
temperature

...
Additional arguments. See find_lw()

Details

Returns a dataframe with all parameters that can be estimated

Examples

fishflux::model_parameters(sp = "Scarus psittacus", family = "Scaridae", temp = 27)

#> average values for b0 and alpha used

#> 
#> SAMPLING FOR MODEL 'get_B0' NOW (CHAIN 1).
#> Chain 1: Iteration:   1 / 1000 [  0%]  (Sampling)
#> Chain 1: Iteration: 100 / 1000 [ 10%]  (Sampling)
#> Chain 1: Iteration: 200 / 1000 [ 20%]  (Sampling)
#> Chain 1: Iteration: 300 / 1000 [ 30%]  (Sampling)
#> Chain 1: Iteration: 400 / 1000 [ 40%]  (Sampling)
#> Chain 1: Iteration: 500 / 1000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0 seconds (Warm-up)
#> Chain 1:                0.00129 seconds (Sampling)
#> Chain 1:                0.00129 seconds (Total)
#> Chain 1: 

#>            species      t0  Linf     k     asp troph   lwa_m      lwa_sd lwb_m
#> 1 Scarus psittacus -0.6025 23.05 0.814 1.80877     2 0.01585 0.001765306  3.06
#>       lwb_sd     mdw_m        f0_m        f0_sd alpha_m   alpha_sd
#> 1 0.02040816 0.2372406 0.001517989 3.134897e-10    0.77 0.05286288