A function to extract specific model output parameters from result

A function to extract specific model output parameters from result

Usage

extract(mod, par)

Arguments

mod

Output from cnp_mod_mcmc()

par

Character vector specifying which output parameter that should be returned.

Value

Main model output parameters:

F0c: C-specific minimal inorganic flux (g/day)

F0n: N-specific minimal inorganic flux (g/day)

F0p: P-specific minimal inorganic flux (g/day)

Gc: Carbon-specific growth rate (g/day)

Gn: Nitrogen-specific growth rate (g/day)

Gp: Phosphorus-specific growth rate (g/day)

Sc: C-specific minimal supply rate (g/day)

Sn: N-specific minimal supply rate (g/day)

Sp: P-specific minimal supply rate (g/day)

Ic: Ingestion rate of C (g/day)

In: Ingestion rate of N (g/day)

Ip: Ingestion rate of P (g/day)

Wc: Egestion rate of C (g/day)

Wn: Egestion rate of N (g/day)

Wp: Egestion rate of P (g/day)

Fc: Total inorganic flux of C (respiration) (g/day)

Fn: Total inorganic flux of N (excretion) (g/day)

Fp: Total inorganic flux of P (excretion) (g/day)

Returns a data.frame with a summary of the selected output parameters

Examples

model <- cnp_model_mcmc(TL = 5:10, param = list(Qc_m = 40, Qn_m = 10, Qp_m = 4))
extract(model, c("Fn","Fp"))