A function to find growth parameters on fishbase

A function to find growth parameters of a species using rfishbase. It returns a data frame containing K, t0 and Linf, the source. This function is useful to see what is available on fishbase. Nevertheless, we strongly recommend to check the source and only use otolith based studies.

growth_params(sp, otolith = TRUE)

Arguments

sp
A charachter value containing the species name

otolith
A logical value. If TRUE, only results from otolith analysis are returned. If false, all growth studies will be returned.

Value

dataframe with available growth rate parameters from fishbase.

Examples

if (FALSE) {
library(fishflux)
growth_params("Lutjanus griseus")
}