A function to calculate metabolic rates

All model parameters below were estimated by Barneche & Allen 2018 Ecology Letters doi: 10.1111/ele.12947. These parameters are for the best model (Model 2 in the paper online supplementary material) of fish resting metabolic rates reported in the paper, which also includes trophic level as a covariate.

Usage

metabolic_rate(temp, troph, asp, B0, m_max, m, a, growth_g_day, f)

Arguments

temp

Temperature in degrees Celsius

troph

Trophic level (from 1 to 5)

asp

The caudal fin aspect ratio , a proxy for activity level

B0

Constant for resting metabolic rate. If NA, function will calculate an average.

m_max

Maximum biomass fish (in g)

m

Wet weight fish (in g)

a

Resting metabolic rate mass-scaling exponent

growth_g_day

Daily growth in grams of wet weight

f

Activity scope (from 1 to 4)

Value

A dataframe with metabolic rates.

Details

All model parameters below were estimated by

Barneche & Allen 2018 Ecology Letters doi: 10.1111/ele.12947. These parameters are for the best model (Model 2 in the paper online supplementary material) of fish resting metabolic rates reported in the paper, which also includes trophic level as a covariate.

Examples

library(fishflux)
fishflux::metabolic_rate(temp = 27, m_max = 600, m = 300, asp = 3,
troph = 2, f = 2, growth_g_day = 0.05, B0 = 0.2, a = 0.6 )
#>   Total_metabolic_rate_C_g_d Total_metabolic_rate_j_d
#> 1                   9.288811                 362263.6
#>   Resting_metabolic_rate_j_d Cost_growth_j_g
#> 1                   181131.8        16.42453