A function to estimate f0 and alpha

All model parameters below were estimated by Barneche & Allen 2018 Ecology Letters doi: 10.1111/ele.12947. These parameters are for the best model (Model 2 in the paper online supplementary material) of fish resting metabolic rates reported in the paper, which also includes trophic level as a covariate.

metabolism(family, temp, troph_m, troph_sd = 1e-10)

Arguments

family
family fish

temp
Temperature in degrees Celsius

troph_m
Trophic level mean (from 1 to 5)

troph_sd
Trophic level sd (optional)

Details

All model parameters below were estimated by Barneche & Allen 2018 Ecology Letters doi: 10.1111/ele.12947. These parameters are for the best model (Model 2 in the paper online supplementary material) of fish resting metabolic rates reported in the paper, which also includes trophic level as a covariate.

Examples

library(fishflux)
metabolism(family = "Pomacentridae", temp = 27, troph_m = 2)

#> values for b0 and alpha on family level

#> 
#> SAMPLING FOR MODEL 'getB0' NOW (CHAIN 1).
#> Chain 1: Iteration:   1 / 1000 [  0%]  (Sampling)
#> Chain 1: Iteration: 100 / 1000 [ 10%]  (Sampling)
#> Chain 1: Iteration: 200 / 1000 [ 20%]  (Sampling)
#> Chain 1: Iteration: 300 / 1000 [ 30%]  (Sampling)
#> Chain 1: Iteration: 400 / 1000 [ 40%]  (Sampling)
#> Chain 1: Iteration: 500 / 1000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0 seconds (Warm-up)
#> Chain 1:                0.001486 seconds (Sampling)
#> Chain 1:                0.001486 seconds (Total)
#> Chain 1: 

#>          f0_m       f0_sd   alpha_m  alpha_sd        b0_m        b0_sd
#> 1 0.004528457 0.002004058 0.4256931 0.1035894 0.001415727 0.0006612888