A function to find a set of parameters

A function to find a set of parameters

Usage

model_parameters(sp, family, otolith = TRUE, temp, ...)

Arguments

sp

Species name

family

family

otolith

TRUE or FALSE, if TRUE, function will only search fishbase for growth parameters that are based upon otolith analysis

temp

temperature

...

Additional arguments to find_lw.

Value

Returns a dataframe with all parameters that can be estimated

Examples

if (FALSE) { # \dontrun{
library(fishflux)
model_parameters(sp = "Scarus psittacus", family = "Scaridae", temp = 27)} # }